Cleaned up some of the logic
- removed versions from dependent resources;
- added create/update/delete for main resource and create_handler/update_handler/delete_handler for dependent resources;
- when instantiating dependent resources, they are searched in etcd first for idempotency;
- added add_finalizers / remove_finalizers to let k8s to handle graceful delete;

+++ containers/mlops/resources/model_deployment.py
@@ -1,4 +1,4 @@
-from typing import Any, Union
+from typing import Any, List, Union
 
 from kubernetes import client as K8SClient
 from pydantic import BaseModel
@@ -6,29 +6,42 @@ from resources.model_storage import ModelStorage
 
 
 class ModelDeployment(BaseModel):
-    name: str
-    namespace: str
-    version: Union[str, None]
-    image: str
-    init_image: str = "quay.io/bdobrica/ml-operator-tools:model-init-latest"
-    artifact: str
-    instances: int
-    cpus: str
-    memory: str
-    deployment: Union[Any, None] = None
+    def __init__(self, name: str, namespace: str):
+        self.name = name
+        self.namespace = namespace
 
-    def get_deployment_body(self) -> K8SClient.V1Deployment:
+        api = K8SClient.AppsV1Api()
+        try:
+            self.deployment = api.read_namespaced_deployment(f"{self.name}-{self.version}", self.namespace)
+        except K8SClient.ApiException as err:
+            if err.status == 404:
+                self.deployment = None
+            else:
+                raise
+
+    def get_deployment_body(
+        self,
+        instances: int,
+        artifact: str,
+        image: str,
+        cpus: str,
+        memory: str,
+        command: List[str] = None,
+        args: List[str] = None,
+        init_image: str = "quay.io/bdobrica/ml-operator-tools:model-init-latest",
+        finalizers: List[str] = None,
+    ) -> K8SClient.V1Deployment:
         deployment_body = K8SClient.V1Deployment(
             metadata=K8SClient.V1ObjectMeta(
-                name=f"{self.name}-{self.version}",
+                name=self.name,
                 namespace=self.namespace,
+                finalizers=finalizers,
             ),
             spec=K8SClient.V1DeploymentSpec(
-                replicas=self.instances,
+                replicas=instances,
                 selector=K8SClient.V1LabelSelector(
                     match_labels={
-                        "app": self.name,
-                        "version": self.version,
+                        "model": self.name,
                     }
                 ),
                 strategy=K8SClient.V1DeploymentStrategy(
@@ -37,19 +50,18 @@ class ModelDeployment(BaseModel):
                 template=K8SClient.V1PodTemplateSpec(
                     metadata=K8SClient.V1ObjectMeta(
                         labels={
-                            "app": self.name,
-                            "version": self.version,
+                            "model": self.name,
                         }
                     ),
                     spec=K8SClient.V1PodSpec(
                         init_containers=[
                             K8SClient.V1Container(
-                                image=self.INIT_IMAGE,
+                                image=init_image,
                                 name=f"{self.name}-init",
                                 env=[
                                     K8SClient.V1EnvVar(
                                         name="MODEL_URL",
-                                        value=self.artifact,
+                                        value=artifact,
                                     ),
                                     K8SClient.V1EnvVar(
                                         name="MODEL_PATH",
@@ -58,7 +70,7 @@ class ModelDeployment(BaseModel):
                                 ],
                                 volume_mounts=[
                                     K8SClient.V1VolumeMount(
-                                        name=f"{self.name}-{self.version}",
+                                        name=self.name,
                                         mount_path="/opt/ml",
                                         read_only=True,
                                     )
@@ -68,20 +80,22 @@ class ModelDeployment(BaseModel):
                         containers=[
                             K8SClient.V1Container(
                                 name=self.name,
-                                image=self.image,
+                                image=image,
+                                command=command,
+                                args=args,
                                 resources=K8SClient.V1ResourceRequirements(
                                     limits={
-                                        "cpu": self.cpus,
-                                        "memory": self.memory,
+                                        "cpu": cpus,
+                                        "memory": memory,
                                     },
                                     requests={
-                                        "cpu": self.cpus,
-                                        "memory": self.memory,
+                                        "cpu": cpus,
+                                        "memory": memory,
                                     },
                                 ),
                                 volume_mounts=[
                                     K8SClient.V1VolumeMount(
-                                        name=f"{self.name}-{self.version}",
+                                        name=self.name,
                                         mount_path="/opt/ml",
                                     ),
                                 ],
@@ -89,7 +103,7 @@ class ModelDeployment(BaseModel):
                         ],
                         volumes=[
                             K8SClient.V1Volume(
-                                name=f"{self.name}-{self.version}",
+                                name=self.name,
                                 persistent_volume_claim=K8SClient.V1PersistentVolumeClaimVolumeSource(
                                     claim_name=f"{self.name}-pvc",
                                 ),
@@ -101,35 +115,78 @@ class ModelDeployment(BaseModel):
         )
         return deployment_body
 
-    def create(self) -> "ModelDeployment":
+    def create(
+        self,
+        instances: int,
+        artifact: str,
+        image: str,
+        cpus: str,
+        memory: str,
+        command: List[str] = None,
+        args: List[str] = None,
+        init_image: str = "quay.io/bdobrica/ml-operator-tools:model-init-latest",
+    ) -> "ModelDeployment":
         if self.deployment is not None:
             return self.update()
 
         api = K8SClient.AppsV1Api()
-        deployment_body = self.get_deployment_body()
+        deployment_body = self.get_deployment_body(
+            instances=instances,
+            artifact=artifact,
+            image=image,
+            cpus=cpus,
+            memory=memory,
+            command=command,
+            args=args,
+            init_image=init_image,
+        )
         self.deployment = api.create_namespaced_deployment(
             namespace=self.namespace,
             body=deployment_body,
         )
         return self
 
-    def update(self) -> "ModelDeployment":
+    def update(
+        self,
+        instances: int,
+        artifact: str,
+        image: str,
+        cpus: str,
+        memory: str,
+        command: List[str] = None,
+        args: List[str] = None,
+        init_image: str = "quay.io/bdobrica/ml-operator-tools:model-init-latest",
+        finalizers: List[str] = None,
+    ) -> "ModelDeployment":
         if self.deployment is None:
             return self.create()
 
         api = K8SClient.AppsV1Api()
-        deployment_body = self.get_deployment_body()
+        deployment_body = self.get_deployment_body(
+            instances=instances,
+            artifact=artifact,
+            image=image,
+            cpus=cpus,
+            memory=memory,
+            command=command,
+            args=args,
+            init_image=init_image,
+            finalizers=finalizers,
+        )
         self.deployment = api.patch_namespaced_deployment(
-            name=f"{self.name}-{self.version}",
+            name=self.name,
             namespace=self.namespace,
             body=deployment_body,
         )
         return self
 
     def delete(self) -> "ModelDeployment":
+        if self.deployment is None:
+            return self
+
         api = K8SClient.AppsV1Api()
         api.delete_namespaced_deployment(
-            name=f"{self.name}-{self.version}",
+            name=self.name,
             namespace=self.namespace,
         )
